Rename component files to match import statements
