+++ app/components/Icons.js
@@ -1,3 +1,4 @@
+/* eslint-disable */
 import React from 'react';
 
 export const IconChannels = (props) => (
