+++ app/components/Loader.js
@@ -1,7 +1,7 @@
 // @flow
 import React from 'react';
-import { IconSpinner } from './icons';
-import styles from './loader.scss';
+import { IconSpinner } from './Icons';
+import styles from './Loader.scss';
 
 const Loader = () => (
   <IconSpinner styles={styles} />
