+++ test/actions/channelList.spec.js
@@ -1,3 +1,4 @@
+/* eslint promise/always-return: 0 */
 import configureMockStore from 'redux-mock-store';
 import thunk from 'redux-thunk';
 import * as channelListActions from '../../app/actions/channelList';
